build: genVersion.mjs 脚本改造
