+++ package.json
@@ -13,7 +13,7 @@
     "test:watch": "jest --watch",
     "gen": "node scripts/createComponent.mjs",
     "gen:icon": "node scripts/genIcons.mjs",
-    "build:version": "node scripts/genVersion.js",
+    "build:version": "node scripts/genVersion.mjs",
     "build:esm": "node scripts/build-esm.js",
     "build:esm-browser": "node scripts/build-browser.js",
     "build:umd": "node scripts/build-umd.js",
@@ -21,7 +21,7 @@
     "build:type": "node scripts/build-types.js",
     "build:icon": "node scripts/build-icon.js",
     "build": "npm run build:version && npm run build:esm && npm run build:type && cp -rf es/icon . && npm run build:esm-browser && npm run build:umd && npm run build:style && npm run build:icon",
-    "release": "node scripts/release.js",
+    "release": "node scripts/release.mjs",
     "lint-staged": "lint-staged --allow-empty",
     "commitlint": "commitlint --config commitlint.config.js -e -V",
     "lint:script": "eslint --ext .js,.ts,.vue,.tsx ./components --fix",
